+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0] - 2023-04-25
+### Added
+- `matches()` and `matchesSubstring()` shortcuts
+
+### Changed
+- More precise interpretation of `.`
+
 ## [0.0.2] - 2023-04-07
 ### Changed
 - API change
@@ -9,5 +16,6 @@
 ### Added
 - Initial release
 
+[0.1.0]: https://github.com/f3ath/iregexp/compare/0.0.2...0.1.0
 [0.0.2]: https://github.com/f3ath/iregexp/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/f3ath/iregexp/releases/tag/0.0.1